# [macOS] Specify kSecUseDataProtectionKeychain when generating RSA/ECC keys