Fixes #36107.
Ref: https://github.com/dotnet/runtime/issues/36107#issuecomment-832278148