@filipnavara 2021-05-14
I'll re-run the benchmarks to make double sure that it takes the right code path.