@filipnavara 2021-05-14
BenchmarkDotNet=v0.12.1.20210514-develop, OS=macOS Big Sur 11.3 (20E232) [Darwin 20.4.0]
Apple M1 2.40GHz, 1 CPU, 8 logical and 8 physical cores
.NET SDK=6.0.100-preview.3.21202.5
  [Host]     : .NET 6.0.0 (6.0.21.20104), X64 RyuJIT
  Job-AAEMFC : .NET 6.0.0 (42.42.42.42424), X64 RyuJIT
  Job-DJNIOL : .NET 6.0.0 (42.42.42.42424), X64 RyuJIT


|     Method |        Job | Toolchain |           Config |        Mean |     Error |    StdDev | Ratio | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |----------- |---------- |----------------- |------------:|----------:|----------:|------:|------:|------:|------:|----------:|
|   SignHash | Job-AAEMFC |    branch | nistP256, SHA256 |    376.8 us |   1.03 us |   0.86 us |  0.07 |     - |     - |     - |     248 B |
|   SignHash | Job-DJNIOL |      main | nistP256, SHA256 |  5,700.9 us |  53.82 us |  47.71 us |  1.00 |     - |     - |     - |     253 B |
|            |            |           |                  |             |           |           |       |       |       |       |           |
| VerifyHash | Job-AAEMFC |    branch | nistP256, SHA256 |    493.8 us |   3.21 us |   2.85 us |  0.04 |     - |     - |     - |   1,337 B |
| VerifyHash | Job-DJNIOL |      main | nistP256, SHA256 | 11,277.0 us | 164.62 us | 137.47 us |  1.00 |     - |     - |     - |   1,341 B |
|            |            |           |                  |             |           |           |       |       |       |       |           |
|   SignHash | Job-AAEMFC |    branch | nistP384, SHA384 |  1,497.6 us |  20.16 us |  17.88 us |  0.13 |     - |     - |     - |     313 B |
|   SignHash | Job-DJNIOL |      main | nistP384, SHA384 | 11,503.5 us | 105.37 us |  93.41 us |  1.00 |     - |     - |     - |     322 B |
|            |            |           |                  |             |           |           |       |       |       |       |           |
| VerifyHash | Job-AAEMFC |    branch | nistP384, SHA384 |  1,524.5 us |  10.95 us |   9.71 us |  0.07 |     - |     - |     - |   1,369 B |
| VerifyHash | Job-DJNIOL |      main | nistP384, SHA384 | 23,298.7 us | 252.31 us | 223.67 us |  1.00 |     - |     - |     - |   1,388 B |
|            |            |           |                  |             |           |           |       |       |       |       |           |
|   SignHash | Job-AAEMFC |    branch | nistP521, SHA512 |  1,540.2 us |  20.36 us |  19.05 us |  0.07 |     - |     - |     - |     393 B |
|   SignHash | Job-DJNIOL |      main | nistP521, SHA512 | 23,319.9 us | 123.77 us | 109.72 us |  1.00 |     - |     - |     - |     404 B |
|            |            |           |                  |             |           |           |       |       |       |       |           |
| VerifyHash | Job-AAEMFC |    branch | nistP521, SHA512 |  1,606.7 us |  13.24 us |  12.39 us |  0.03 |     - |     - |     - |   1,409 B |
| VerifyHash | Job-DJNIOL |      main | nistP521, SHA512 | 46,410.1 us | 150.48 us | 117.48 us |  1.00 |     - |     - |     - |   1,465 B |