@filipnavara 2021-05-15
# RSA key generation benchmark

BenchmarkDotNet=v0.12.1.20210515-develop, OS=macOS Big Sur 11.3 (20E232) [Darwin 20.4.0]
Apple M1 2.40GHz, 1 CPU, 8 logical and 8 physical cores
.NET SDK=6.0.100-preview.3.21202.5
  [Host]     : .NET 6.0.0 (6.0.21.20104), X64 RyuJIT
  Job-VMVYPG : .NET 6.0.0 (42.42.42.42424), X64 RyuJIT
  Job-FZFMXK : .NET 6.0.0 (42.42.42.42424), X64 RyuJIT


|      Method |        Job | Toolchain | KeySize |        Mean |      Error |       StdDev |      Median | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |----------- |---------- |-------- |------------:|-----------:|-------------:|------------:|------:|--------:|------:|------:|------:|----------:|
| GenerateKey | Job-VMVYPG |    branch |    1024 |    22.68 ms |   0.618 ms |     1.813 ms |    22.62 ms |  0.47 |    0.09 |     - |     - |     - |      4 KB |
| GenerateKey | Job-FZFMXK |      main |    1024 |    49.74 ms |   3.082 ms |     9.038 ms |    49.04 ms |  1.00 |    0.00 |     - |     - |     - |      4 KB |
|             |            |           |         |             |            |              |             |       |         |       |       |       |           |
| GenerateKey | Job-VMVYPG |    branch |    2048 |   196.33 ms |  21.201 ms |    61.846 ms |   185.20 ms |  0.66 |    0.49 |     - |     - |     - |      6 KB |
| GenerateKey | Job-FZFMXK |      main |    2048 |   427.20 ms |  83.391 ms |   241.932 ms |   392.75 ms |  1.00 |    0.00 |     - |     - |     - |      7 KB |
|             |            |           |         |             |            |              |             |       |         |       |       |       |           |
| GenerateKey | Job-VMVYPG |    branch |    4096 | 2,437.88 ms | 529.845 ms | 1,537.176 ms | 1,974.58 ms |  0.90 |    1.43 |     - |     - |     - |     12 KB |
| GenerateKey | Job-FZFMXK |      main |    4096 | 4,527.22 ms | 894.447 ms | 2,537.396 ms | 4,167.44 ms |  1.00 |    0.00 |     - |     - |     - |     12 KB |