@filipnavara 2021-05-17
The problem with MacCatalyst / iOSSimulator should be fixed in `main`. Re-running the tests should succeeded.